Ajeitando as nomenclaturas
