+++ src/app.jsx
@@ -1,14 +1,14 @@
 import React from "react";
 import { BrowserRouter, Routes, Route} from "react-router-dom";
 import { Login } from "./pages/Login";
-import {Cadastro} from "./pages/Cadastro"
+import { Register } from "./pages/Register"
 
 export function App() {
   return (
     <BrowserRouter>
         <Routes>
             <Route path="/" element={<Login/>}/>
-            <Route path="/Cadastro" element={<Cadastro/>}/>
+            <Route path="/Register" element={<Register/>}/>
             <Route path="/questoes" element={<Login/>}/>
             <Route path="/turmas" element={<Login/>}/>
             <Route path="/provas" element={<Login/>}/>
