+++ src/pages/Register/index.jsx
@@ -1,9 +1,44 @@
 import React from "react";
-import { Header } from "../../components/Header";
-import { Footer } from "../../components/Footer";
+import Button from "../../components/Button";
+import Input from "../../components/Input";
+import { Container } from "./styles";
+import { useState } from 'react';
 
 export function Register() {
+
+    /*const [userType, setUserType] = useState([false,false]);*/
+
+    /*const handleOnChange = (position) => {
+        const updatedCheckedState = user.map((item, index)
+        index === position ? true : false
+        );
+
+        setUserType(updatedCheckedState);
+    };*/
+
     return (
-        <><Header /><h1>Page Register </h1><Footer /></>
+        <Container>
+            <h1>Criar conta</h1>
+            <p>Nome</p>
+            <Input/>
+            <p>Email</p>
+            <Input/>
+            <p>Senha</p>
+            <Input/>
+            <div className="checkbox__container">
+                <div className="checkboxAndUser">
+                    <input id='checkboxStudent' type="checkbox" /*checked={userType[0]}
+                    onChange{() => handleOnChange(0)} */></input>
+                    <label for="checkboxTeacher">Aluno</label>
+                </div>
+                <div className="checkboxAndUser">
+                    <input id="checkboxTeacher" type="checkbox"/*checked={userType[1]}
+                    onChange{() => handleOnChange(1)} */></input>
+                    <label for="checkboxTeacher">Professor</label>
+                </div>
+            </div>
+            <Button> Cadastrar </Button>
+            <Button url="/"> Voltar à página de Login </Button>
+        </Container>
     )
 }
